refactor:修改shopify和walmart token模块

+++ yudao-module-oms/yudao-module-oms-biz/src/main/java/cn/iocoder/yudao/module/oms/model/entity/OmsShop.java
@@ -96,7 +96,7 @@ public class OmsShop {
      * token表主键
      */
     @TableField(value = "auth_id")
-    private Integer authId;
+    private Long authId;
 
     /**
      * 原始json
